requirements.txt
transformers
pytorch-crf
nltk
spacy
